[Inspection Service] Clean up crate layout.
