+++ aptos-node/src/services.rs
@@ -21,7 +21,7 @@ use aptos_storage_interface::{DbReader, DbReaderWriter};
 use aptos_time_service::TimeService;
 use aptos_types::chain_id::ChainId;
 use futures::channel::{mpsc, mpsc::Sender};
-use std::{sync::Arc, thread, time::Instant};
+use std::{sync::Arc, time::Instant};
 use tokio::runtime::Runtime;
 
 const AC_SMP_CHANNEL_BUFFER_SIZE: usize = 1_024;
@@ -134,13 +134,7 @@ pub fn start_node_inspection_service(
     node_config: &NodeConfig,
     peers_and_metadata: Arc<PeersAndMetadata>,
 ) {
-    let node_config = node_config.clone();
-    thread::spawn(move || {
-        aptos_inspection_service::inspection_service::start_inspection_service(
-            node_config,
-            peers_and_metadata,
-        )
-    });
+    aptos_inspection_service::start_inspection_service(node_config.clone(), peers_and_metadata)
 }
 
 /// Starts the peer monitoring service and returns the runtime
