+++ crates/aptos-inspection-service/src/inspection_client.rs
@@ -5,11 +5,7 @@ use anyhow::Result;
 use reqwest::Url;
 use std::collections::HashMap;
 
-pub struct InspectionClient {
-    client: reqwest::Client,
-    url: Url,
-}
-
+/// A simple metric value enum (to represent different value types)
 #[derive(Clone, Debug)]
 pub enum MetricValue {
     I64(i64),
@@ -18,6 +14,7 @@ pub enum MetricValue {
 }
 
 impl MetricValue {
+    /// Convert the value to i64
     pub fn to_i64(&self) -> Result<i64> {
         match self {
             MetricValue::I64(v) => Ok(*v),
@@ -26,6 +23,7 @@ impl MetricValue {
         }
     }
 
+    /// Convert the value to f64
     pub fn to_f64(&self) -> Result<f64> {
         match self {
             MetricValue::I64(v) => Err(anyhow::format_err!("Value not f64: {}", v)),
@@ -35,39 +33,36 @@ impl MetricValue {
     }
 }
 
+/// A simple inspection client for querying metrics from a node
+pub struct InspectionClient {
+    client: reqwest::Client,
+    url: Url,
+}
+
 impl InspectionClient {
-    /// Create an InspectionClient from a valid socket address
-    pub fn new<A: AsRef<str>>(client: reqwest::Client, address: A, port: u16) -> Self {
-        let addr = format!("http://{}:{}", address.as_ref(), port);
-
-        Self {
-            client,
-            url: Url::parse(&addr).unwrap(),
-        }
-    }
-
-    pub fn from_url(url: Url) -> Self {
+    /// Create a new client from the given url
+    pub fn new(url: Url) -> Self {
         let client = reqwest::Client::new();
-
         Self { client, url }
     }
 
+    /// Get an i64 metric value from the node
     pub async fn get_node_metric_i64<S: AsRef<str>>(&self, metric: S) -> Result<Option<i64>> {
-        let metrics = self.get_node_metrics().await?;
-        metrics
+        let node_metrics = self.get_forge_metrics().await?;
+        node_metrics
             .get(metric.as_ref())
             .map_or(Ok(None), |v| v.to_i64().map(Some))
     }
 
-    /// Retrieves all node metrics for a given metric name.
-    /// Allows for filtering metrics by fields afterwards.
+    /// Retrieves all node metrics for a given metric name
     pub async fn get_node_metric_with_name(
         &self,
-        metric: &str,
+        metric_name: &str,
     ) -> Result<Option<HashMap<String, MetricValue>>> {
-        let metrics = self.get_node_metrics().await?;
-        let search_string = format!("{}{{", metric);
+        let metrics = self.get_forge_metrics().await?;
+        let search_string = format!("{}{{", metric_name);
 
+        // Filter out all metrics that don't start with the search string
         let result: HashMap<_, _> = metrics
             .iter()
             .filter_map(|(key, value)| {
@@ -79,6 +74,7 @@ impl InspectionClient {
             })
             .collect();
 
+        // Return None if the result is empty
         if result.is_empty() {
             Ok(None)
         } else {
@@ -86,11 +82,13 @@ impl InspectionClient {
         }
     }
 
-    pub async fn get_node_metrics(&self) -> Result<HashMap<String, MetricValue>> {
+    /// Fetches and returns all node metrics by pinging the forge_metrics endpoint
+    pub async fn get_forge_metrics(&self) -> Result<HashMap<String, MetricValue>> {
         let mut url = self.url.clone();
         url.set_path("forge_metrics");
-        let response = self.client.get(url).send().await?;
 
+        // Fetch the metrics from the node
+        let response = self.client.get(url).send().await?;
         response
             .json::<HashMap<String, String>>()
             .await?
